requirements.txt
Django==5.1.7
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pillow==11.1.0
openpyxl==3.1.5
matplotlib==3.10.1
numpy==2.2.4
xhtml2pdf==0.2.11
requests==2.32.5
python-dotenv==1.2.1